feat: support stable web read images tools
